requirements.txt
httpx>=0.27
pyyaml>=6.0
openai>=1.0
